Use 16-bit floating point on the GPU instead of 8-bit integers
Since cuDNN does not seem to have an implementation of convolution that
uses the tensor cores for integers (yet?), use 16-bit floating points
instead. This is should be both faster, and stronger since it is more
accurate.

8-bit integers
---

```
test batch_size_008 ... bench:   1,045,368 ns/iter (+/- 6,896)
test batch_size_016 ... bench:   1,830,196 ns/iter (+/- 6,412)
test batch_size_032 ... bench:   2,712,810 ns/iter (+/- 41,591)
```

16-bit floating points
---

```
test batch_size_008 ... bench:     837,513 ns/iter (+/- 8,251)
test batch_size_016 ... bench:   1,316,623 ns/iter (+/- 9,070)
test batch_size_032 ... bench:   2,066,762 ns/iter (+/- 15,927)
```

+++ src/mcts/predict.rs
@@ -18,6 +18,7 @@ use parallel::{self, OneSender};
 use go::util::features::{FEATURE_SIZE};
 use nn::devices::{DEVICES, set_current_device};
 use nn::{self, Network, Output, OutputSet, Workspace};
+use util::types::f16;
 use util::config;
 
 pub type PredictGuard<'a> = parallel::ServiceGuard<'a, PredictState>;
@@ -29,7 +30,7 @@ pub fn service(network: Network) -> PredictService {
 
 pub enum PredictRequest {
     /// Request to compute the value and policy for some feature.
-    Ask(Vec<i8>),
+    Ask(Vec<f16>),
 
     /// Indicate that a worker is waiting for some other thread to finish
     /// and should be awaken after the next batch of computations finish.
@@ -45,7 +46,7 @@ pub struct PredictState {
     running_count: AtomicUsize,
 
     /// The features to get the value and policy for.
-    features_list: Vec<i8>,
+    features_list: Vec<f16>,
 
     /// The sender to response to each of the features in `features_list`
     /// over.
@@ -80,7 +81,7 @@ impl PredictState {
     /// * `workspace` - 
     /// * `features_list` - 
     /// 
-    fn forward(workspace: &mut Workspace, features_list: &[i8]) -> (Vec<f32>, Vec<Vec<f32>>) {
+    fn forward(workspace: &mut Workspace, features_list: &[f16]) -> (Vec<f32>, Vec<Vec<f32>>) {
         let mut outputs = nn::forward(
             workspace,
             features_list,
